Server/requirements.txt
Flask==2.3.0
Flask-SQLAlchemy==3.0.5
Flask-Login==0.6.2
python-libsignal==0.20.0
cryptography==40.0.2
qrcode==7.4.2
requests==2.30.0
python-dotenv==1.0.0
PyJWT==2.6.0
Flask-CORS==4.0.0
gunicorn==20.1.0
psycopg2-binary==2.9.6
GitPython==3.1.31
Werkzeug==2.3.0
